Add function to store kernel config after running core.sh
This will store the kernelname config inside the folder resources/core-devices/
and everytime you run core.sh if you enter the same kernel name the program will
ask to load that kernel config introduced before.

The goal of this is to save time when you run core.sh and configure the same kernel.