KB-E: Re-Write Devices data saving method and all scripts related
Before, KB-E after settings everything up would have saved the
device configuration and values under the "kernelname", inside
./devices/<kernelname>/ folder, this was the first attempt to
make KB-E to store and load devices without going through the
configuration process, the command "kbe start" usage was:

$ kbe start <kernelname>

This was inefficient because if you wanted to make a kernel for
another device under the same kernelname, it would override the
old device settings.

Now, after the user runs KB-E and all settings and values are
set, a ./devices/<device>/ folder gets created and inside that
device folder a ./devices/<device>/<kernelname> folder is created
where all your settings and values are stored. Take in consideration
that the <device> that we're talking about is the same "device
codename" that KB-E asks for when you go through the config process.
The new usage for the command "kbe start" is:

$ kbe start <device> <kernelname>
                      (optional)

How does this works now? If you run "kbe start" followed by a <device>
you configured before, the program will check for available kernels
configurations at it's ./devices/<device>/ folder. There are four
things to take in consideration:

1- If you didn't specified the <kernelname>, KB-E will determine if
for that device theres one or more kernel configurations.

2- If theres only one kernel configuration, KB-E will automatically
load it and be ready for its use

3- If theres more than one kernel configuration, KB-E will list you
all of them and make you to select one, then, it will be ready
for its use

4- If you specify the <kernelname> on a device that has more than
one kernel configuration, KB-E will skip the list selection, load it
and be ready for its use

This is it, now KB-E support multiple kernel configuration or
"variants" of the same device, this will make up for the MultiVariants
support I removed.