KB-E: Add option to build DTB Manually
This will add to new devices and existing ones (via updates.sh,
please run "kbe upgrade") an option to manually build DTB, kbe
"--dtb & -dt" flags will be soon removed and DTB building will
be automatic depending on this new option (for the next commit).

This is because some kernel sources on build completed there can
be a kernel Image with the dtb already appended to it, so, there
wouldn't be need for KB-E to build it.

Rev bump:00-->10