buildkernel.sh: Introduce function to automatically select the appropiate Kernel Image
This function "selectimage" in short words automatically selects the appropiate
kernel image for the user, the name of the kernel that it selects is stored in
the variable "selected_image", this function will always select an image no matter
what except only if the kernel is not built. For more information read the code,
it's full of info.