KB-E: Introduce "buildkernel-assistant.sh" script
This scripts handles the kernel building output by the crosscompiler
in a more "fancy" way, by redirecting the CC output to a log.txt file,
at the same time tailing each output and displaying a more "clean"
information to the user.

I'll add a compile error assistant to give basic help in case of an
error compiling the kernel, soon.