buildkernel.sh: Restore Kernel Debug functionality
Before the introduction of buildkernel-assistant.sh, when you
enabled the kernel debugging during the device configuration,
buildkernel.sh would show you the compile process, if it were
disabled, it shows nothing. Taking the choice of not "debug"
the compile process were not good at all because there was
nothing telling you that it was doing something.

Now, with the new assistant script, there is a more clean
compiling process output for the user, so, with this commit,
decidingnot to "debug" the compile process will show you the
new fancy compiling process by buildkernel-assistant.sh, and,
enabling it, will show you instead the original crosscompiler
output.