KernelBuilding-Essentials Core Update
KB-E is in process of beign "re-written", these and upcoming
changes will be uploaded to a kbe-beta branch because it
modifies almost everything from functions to variables and how
in general the program works. the main script "core.sh" has
been renamed to "kbe.sh" and re-written.

This new KernelBuilding Essentials will be version 1.0-Beta
and when everything is fixed and updated, I will write a
proper README.md with all the info necessary and another
README.md for the modules/ folder with all the info about
KB-E modules system, commands and variables with their
descriptions.

I do not recommend to download and use this branch of KB-E
because it's still Beta and I've various things to update and
fix. I'd rather recommend to wait for the kbe-stable branch
to be released soon.