programtools, runsettings: Move to Linaro-GCC
I was thinking of adding suppport for Clang for default to KB-E,
but after reading some articles and looking at benchmarks, I came
to the conclusion that it is not worth the effort, instead, because
google will deprecate it's GCC 4.9 completely in 2020, let's use
Linaro by default.

Also, letting users set a custom ToolChain is something I've taken
in consideration all this time, so, I'll make a way for this soon.